skinflow-emr-v4/requirements.txt
Django==5.2.9
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
requests
Pillow
django-storages[s3]
boto3
weasyprint

